Restart clean
