+++ .quarkus/cli/plugins/quarkus-cli-catalog.json
@@ -1,0 +1,13 @@
+{
+  "version" : "v1",
+  "lastUpdate" : "26/12/2024 22:31:30",
+  "plugins" : {
+    "tls" : {
+      "name" : "tls",
+      "type" : "maven",
+      "location" : "io.quarkus:quarkus-tls-registry-cli:3.17.5",
+      "description" : null,
+      "inProjectCatalog" : false
+    }
+  }
+}
