+++ src/main/resources/application.properties
@@ -1,5 +1,1 @@
 quarkus.http.port=8081
-
-quarkus.datasource.db-kind=postgresql
-quarkus.datasource.jdbc=false
-quarkus.flyway.migrate-at-start=true
