--- src/test/resources/application.properties
@@ -1,2 +1,0 @@
-quarkus.flyway.clean-at-start=true
-quarkus.hibernate-orm.log.sql=true
