--- src/test/resources/data.sql
@@ -1,1 +1,0 @@
-INSERT INTO zikresources (id, url, title) VALUES ('24705f60-7f2a-491b-a5d4-16467662cdd2', 'https://www.guitare6.fr/tabs-pdf/tabs.php?pdf=JimiHendrix/LittleWing-1', 'Little wing');
