--- src/test/resources/import.sql
@@ -1,4 +1,0 @@
-INSERT INTO zikresources(_id, url, title, artist, tags) 
-VALUES ('c5d027da-1689-4217-8309-edd8283e99fe', 'https://www.songsterr.com/a/wsa/tool-sober-tab-s19923t2', 'Sober', 'Tool', null);
-INSERT INTO zikresources(_id, url, title, artist, tags) 
-VALUES ('e337298e-eec1-4443-b0b7-77e9f307dec9', 'https://www.songsterr.com/a/wsa/pouet', 'Test to delete', 'Erwann and Manu', null);
